requirements.txt
flake8>=3.7.9
matplotlib>=3.1.2
numpy>=1.18.0
pandas>=1.0.3
pytest>=5.4.2
pyyaml>=5.3.1
scipy>=1.3.1
torch>=1.3.1
